test: Add BaseUITest class
Add a base class for UITests for the iOSSwift to remove
some redundant code.
